fixed file claim, added refresesh feature to add and update claim

+++ TestJPA/src/main/java/org/example/app/components/buttonSet/MyDependantButtonSet.java
@@ -1,4 +1,4 @@
-package org.example.app.components.buttonSet;
+package org.example.app.components.buttonset;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -7,6 +7,7 @@ import javafx.scene.control.Button;
 import javafx.scene.layout.HBox;
 import org.example.app.components.form.AddDependantForm;
 import org.example.app.components.table.DependantTable;
+import org.example.app.controllers.CustomerAdminController;
 import org.example.model.customer.Dependant;
 import org.example.repository.impl.CustomerRepository;
 
@@ -24,8 +25,11 @@ public class MyDependantButtonSet extends HBox {
     @FXML
     private Button setPolicyHolderButton;
     private DependantTable dependantTable;
-    public MyDependantButtonSet(DependantTable dependantTable) {
+    private CustomerAdminController controller;
+
+    public MyDependantButtonSet(DependantTable dependantTable, CustomerAdminController controller) {
         this.dependantTable = dependantTable;
+        this.controller = controller;
         FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/views/components/dependantButtonSet.fxml"));
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
@@ -35,6 +39,7 @@ public class MyDependantButtonSet extends HBox {
         } catch (IOException exception) {
             throw new RuntimeException(exception);
         }
+        setUpButtonSet();
     }
 
     private void setUpButtonSet() {
@@ -50,15 +55,17 @@ public class MyDependantButtonSet extends HBox {
     private void update(ActionEvent actionEvent) {
 
     }
-
     private void delete(ActionEvent actionEvent) {
         CustomerRepository repository = new CustomerRepository();
         Dependant dependant = dependantTable.getSelectionModel().getSelectedItem();
-        repository.removeByID(dependant.getId());
-        repository.close();
+        if (dependant != null) {
+            repository.removeByID(dependant.getId());
+            dependantTable.getItems().remove(dependant);
+            repository.close();
+        }
     }
 
     private void add(ActionEvent actionEvent) {
-        new AddDependantForm();
+        new AddDependantForm(controller);  // Now correctly passes the controller
     }
 }
